Update pdfstitcher.po
Hello, 

Here is a first version of the French translation. 

On line 151-153 #: pdfstitcher.py:198
msgid "Exception"
msgstr ""
I haven't translated this one yet, I would need more context as to where and when this message appears. I have tried the software with lots of files and options but haven't encountered this one.

Also, I had an error [errno 13] permission denied, but you didn't include the error messages in the strings, would you like to add them?

Finally, I have tried to be accurate and compact in the translations, but I dont know what you'll be doing regarding text boxes and buttons sizes etc when the translations don't fit. So don't hesitate to tell me if some are too long. I am a translator specialized in localization, so translating softwares I know that kind of issues can be time consuming. Let me know if I can help on that front.

Thank you for your great job, this Stitcher is amazing!
Have a nice day!

Marie